+++ archive/nerves_bootstrap/lib/mix/tasks/nerves.loadpaths.ex
@@ -1,5 +1,7 @@
 defmodule Mix.Tasks.Nerves.Loadpaths do
   use Mix.Task
+
+  import Mix.Nerves.Bootstrap.Utils
   alias Nerves.Env
 
   def run(_args) do
@@ -9,16 +11,16 @@ defmodule Mix.Tasks.Nerves.Loadpaths do
           Env.initialize
           try do
             Env.bootstrap
-            Mix.shell.info "Nerves Env loaded"
-            shell_info
+            debug_info "Nerves Env loaded"
+            env_info
           rescue
             UndefinedFunctionError ->
-              Mix.shell.info "Nerves Env needs to be updated"
+              debug_info "Nerves Env needs to be updated"
             e ->
               raise e
           end
         false ->
-          Mix.shell.info "Nerves Env not loaded"
+          debug_info "Nerves Env not loaded"
       end
     end
   end
@@ -29,17 +31,15 @@ defmodule Mix.Tasks.Nerves.Loadpaths do
     |> Path.relative_to_cwd
   end
 
-  def shell_info do
-    if System.get_env("NERVES_DEBUG") == "1" do
-      Mix.shell.info """
-      ------------------
-      Nerves Environment
-      ------------------
-      target:     #{Mix.Project.config[:target]}
-      toolchain:  #{env("NERVES_TOOLCHAIN")}
-      system:     #{env("NERVES_SYSTEM")}
-      app:        #{env("NERVES_APP")}
-      """
-    end
+  def env_info do
+    debug_info """
+    ------------------
+    Nerves Environment
+    ------------------
+    target:     #{Mix.Project.config[:target]}
+    toolchain:  #{env("NERVES_TOOLCHAIN")}
+    system:     #{env("NERVES_SYSTEM")}
+    app:        #{env("NERVES_APP")}
+    """
   end
 end