+++ archive/nerves_bootstrap/templates/new/lib/application_name.ex
@@ -4,7 +4,7 @@ defmodule <%= application_module %> do
   # See http://elixir-lang.org/docs/stable/elixir/Application.html
   # for more information on OTP Applications
   def start(_type, _args) do
-    import Supervisor.Spec
+    import Supervisor.Spec, warn: false
 
     # Define workers and child supervisors to be supervised
     children = [