+++ lib/mix/nerves/utils.ex
@@ -1,15 +1,14 @@
 defmodule Mix.Nerves.Utils do
+  @fwup_semver "~> 0.8"
 
-  def shell(cmd) do
+  def shell(cmd, args) do
     stream = IO.binstream(:standard_io, :line)
-    Application.put_env(:porcelain, :driver, Porcelain.Driver.Basic)
-    Application.ensure_started(:porcelain)
-    Porcelain.shell(cmd, in: stream, async_in: true, out: stream, err: :out)
+    System.cmd(cmd, args, into: stream, stderr_to_stdout: true)
   end
 
   def preflight do
     check_requirements
-    Mix.Task.run "nerves.loadpaths", []
+    Mix.Task.run "nerves.loadpaths"
   end
 
   def check_requirements do
@@ -22,8 +21,18 @@ defmodule Mix.Nerves.Utils do
       """
     end
 
-    case System.cmd("which", ["fwup"]) do
-      {_, 0} -> nil
+    case System.cmd("fwup", ["--version"]) do
+      {vsn, 0} ->
+        vsn = String.strip(vsn)
+        {:ok, req} = Version.parse_requirement(@fwup_semver)
+        unless Version.match?(vsn, req) do
+          Mix.raise """
+          fwup #{@fwup_semver} is required for Nerves.
+          You are running #{vsn}.
+          Please see https://hexdocs.pm/nerves/installation.html#fwup
+          for installation instructions
+          """
+        end
       _ -> Mix.raise """
       fwup is required to create and burn firmware.
       Please see https://hexdocs.pm/nerves/installation.html#fwup
@@ -46,4 +55,10 @@ defmodule Mix.Nerves.Utils do
     end
   end
   def check_host_requirements(_), do: nil
+
+  def debug_info(msg) do
+    if System.get_env("NERVES_DEBUG") == "1" do
+      Mix.shell.info(msg)
+    end
+  end
 end