+++ archive/nerves_bootstrap/CHANGELOG.md
@@ -1,5 +1,12 @@
 # nerves_bootstrap
 
+## v0.1.4
+  * Bug Fixes
+    * Do not warn on import Supervisor.Spec
+    * Silence alias location messages unless NERVES_DEBUG=1
+  * Enhancements
+    * Support for Elixir 1.3.2
+
 ## v0.1.3
   * Enhancements
     * Support for elixir ~> 1.3.0-rc.0